+++ app/Http/Requests/Projects/NewProjectApp.php
@@ -27,7 +27,8 @@ class NewProjectApp extends FormRequest
             'provider' => 'required|in:github,bitbucket',
             'repository' => 'required|nullable|regex:_^([a-z-]+)/([a-z-]+)$_i',
             'branch' => 'nullable|string',
-            'config' => 'sometimes|required|array',
+            'config' => 'sometimes|required|array:phpVersion',
+            'config.phpVersion' => 'sometimes|required|regex:/^[7-8]\.[0-4]$/',
         ];
     }