+++ app/Projects/Application.php
@@ -4,7 +4,7 @@ namespace Servidor\Projects;
 
 use Exception;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Casts\AsArrayObject;
+use Illuminate\Database\Eloquent\Casts\AsCollection;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Support\Carbon;
@@ -66,7 +66,7 @@ class Application extends Model
     ];
 
     protected $casts = [
-        'config' => AsArrayObject::class,
+        'config' => AsCollection::class,
     ];
 
     protected $dispatchesEvents = [